@@ -7,7 +7,7 @@ import (
 	"math"
 )
 
-// TODO: only considering spell's first rank atm, but need to consider all (e.g. 'q' has 5 ranks, etc.)
+// TODO: only considering spell's max rank atm (i.e. spell.MaxRank-1), but need to consider all (e.g. 'q' has 5 ranks, etc.)
 
 type Tactics interface {
 	Fight(champion1, champion2 Champion)
@@ -22,14 +22,14 @@ func New(log log.Logger) Tactics {
 }
 
 type bestSolution struct {
-	Benchmark     float32 // time (in seconds) taken to slay the enemy
+	Benchmark     float64 // time (in seconds) taken to slay the enemy
 	RoundOfSpells []Spell
 }
 
 // Fight Champion1 vs Champion2 health point
 func (f *FightTactics) Fight(champion1, champion2 Champion) {
 	var sol []Spell
-	var bestSol = bestSolution{Benchmark: math.MaxFloat32, RoundOfSpells: []Spell{}}
+	var bestSol = bestSolution{Benchmark: math.MaxFloat64, RoundOfSpells: []Spell{}}
 
 	getBestRoundOfSpells(0, champion1.Spells, sol, champion2.Stats.Hp, &bestSol)
 
@@ -40,7 +40,7 @@ func (f *FightTactics) Fight(champion1, champion2 Champion) {
 	file.Write(fileName, getRoundSpellsToString(bestSol.RoundOfSpells, champion2.Stats.Hp, bestSol.Benchmark))
 }
 
-func getBestRoundOfSpells(pos int, spells, sol []Spell, hp float32, bestSol *bestSolution) {
+func getBestRoundOfSpells(pos int, spells, sol []Spell, hp float64, bestSol *bestSolution) {
 	if isHpZero(sol, hp) {
 		setBenchmark(sol, bestSol)
 		return
@@ -57,7 +57,7 @@ func getBestRoundOfSpells(pos int, spells, sol []Spell, hp float32, bestSol *bes
 }
 
 // isHpZero True if hp is zero, false otherwise
-func isHpZero(sol []Spell, hp float32) bool {
+func isHpZero(sol []Spell, hp float64) bool {
 	for _, spell := range sol {
 		hp = hp - spell.Damage[spell.MaxRank-1]
 		if hp <= 0 {
@@ -71,39 +71,36 @@ func setBenchmark(spells []Spell, bestSol *bestSolution) {
 	tmpBench := getBenchmark(spells)
 	if tmpBench < bestSol.Benchmark {
 		bestSol.Benchmark = tmpBench
+
+		bestSol.RoundOfSpells = make([]Spell, len(spells))
 		copy(bestSol.RoundOfSpells, spells)
 	}
 }
 
 // getBenchmark Given a set of spells (which brings the hp to zero), return the relevant benchmark (i.e. time needed to kill the enemy)
-func getBenchmark(spells []Spell) float32 {
-	var benchmark float32
-	var usedSpells []Spell
-
-	for _, spell := range spells {
-		benchmark += spell.Cast + getAdditionalTimeIfSpellIsInCooldown(spell, usedSpells)
-		usedSpells = append(usedSpells, spell)
+func getBenchmark(spells []Spell) (benchmark float64) {
+	usedSpells := make([]Spell, len(spells))
+	for i, s := range spells {
+		benchmark += s.Cast + getAdditionalTimeIfSpellIsInCooldown(s, usedSpells)
+		usedSpells[i] = s
 	}
-
 	return benchmark
 }
 
 // getAdditionalTimeIfSpellIsInCooldown Get additional waiting time if the spell has been used previously and is therefore still in cooldown.
-func getAdditionalTimeIfSpellIsInCooldown(currentSpell Spell, usedSpells []Spell) float32 {
-	var usedSpellPos = -1
-	var timeToWait float32
-
-	// Check if spell has been used already
-	for i, usedSpell := range usedSpells {
-		if currentSpell.ID == usedSpell.ID {
-			usedSpellPos = i + 1 // (+1 to exclude this spell from the count below, i.e. cooldown starts after cast)
+func getAdditionalTimeIfSpellIsInCooldown(currentSpell Spell, usedSpells []Spell) (timeToWait float64) {
+	// Check last time spell was used
+	usedSpellPos := -1
+	for i := len(usedSpells) - 1; i >= 0; i-- {
+		if currentSpell.ID == usedSpells[i].ID {
+			usedSpellPos = i + 1 // +1 to exclude this spell from the count below, i.e. cooldown starts after cast
 			break
 		}
 	}
 
 	// If it has been used before, check if spell is still in cooldown or not
 	if usedSpellPos != -1 {
-		var timePassed float32
+		var timePassed float64
 
 		// Compute how much time has passed between the first usage of the spell and its re-usage
 		for i := usedSpellPos; i < len(usedSpells); i++ {
@@ -123,9 +120,7 @@ func getAdditionalTimeIfSpellIsInCooldown(currentSpell Spell, usedSpells []Spell
 }
 
 // getRoundSpellsToString Format spells into string
-func getRoundSpellsToString(spells []Spell, hp, benchmark float32) string {
-	var spellsToString string
-
+func getRoundSpellsToString(spells []Spell, hp, benchmark float64) (spellsToString string) {
 	for _, s := range spells {
 		spellsToString += fmt.Sprintf("%s: %.2f (hp: %.2f -> %.2f)\n", s.ID, s.Damage, hp, hp-s.Damage[s.MaxRank-1])
 		hp = hp - s.Damage[s.MaxRank-1]
